LayoutTests/webgpu/map-read-buffers-expected.txt

CONSOLE MESSAGE: GPUBuffer.mapReadAsync(): Invalid operation!
CONSOLE MESSAGE: GPUBuffer.mapWriteAsync(): Invalid operation!
CONSOLE MESSAGE: GPUBuffer.mapReadAsync(): Invalid operation!
CONSOLE MESSAGE: GPUBuffer.mapReadAsync(): Invalid operation!

PASS mapReadAsync, unmap, and destroy on a GPUBuffer.
PASS GPUBuffer.mapReadAsync on a buffer created via GPUDevice.createBufferMapped.
PASS Reject a map read on a buffer not created with MAP_READ usage.
PASS Reject a map read on a mapped GPUBuffer.
PASS Reject a pending map read if GPUBuffer is unmapped.
